backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
fastapi==0.115.4
fonttools==4.54.1
geographiclib==2.0
geojson==3.1.0
geopandas==1.0.1
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
osmium==4.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
